LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/broadcastchannel-success.https-expected.txt


FAIL Structured cloning of SharedArrayBuffers: BroadcastChannel within the same agent cluster assert_equals: The first iframe must have set view[0] to 1 expected 1 but got 0
